+++ src/kernel/drivers/ErrorHandling/Exceptions.c
@@ -1,16 +1,34 @@
 #include "ErrorHandling/Exceptions.h"
 
 void HandlePageFault(InterruptRegistersError* regs){
-    printf("Page fault\n");
-
     PageEntries newPhysical = FindNextFreePhysical();
     uint64_t virtual_address = regs->cr2;
 
-    void* confirmation;
+    SetTextColor(LRED, BLACK);
 
-    if(virtual_address > VIRTUAL_MEMORY_BARRIER){
-        printf("User\n");
-        
+    uint64_t CR3;
+    asm volatile(
+        "movq %%cr3, %%rax\n"
+        "movq %%rax, %0\n"
+        : "=r" (CR3) : :
+    );
+
+    #ifdef DEBUG
+        printf("(!) Page fault\n\tCR2: %x\n\tERR: %b\n\tRIP: %x\n\tCR3: %x\n\tFrom: ", virtual_address, regs->error_code, regs->rip, CR3);
+   
+        if(regs->rip < VIRTUAL_MEMORY_BARRIER){
+            printf("User | For: ");
+        }else{
+            printf("Kernel | For: ");
+        }
+    #endif
+
+    virtual_address = virtual_address & ~0xFFF;
+
+    if(virtual_address < VIRTUAL_MEMORY_BARRIER){
+        #ifdef DEBUG
+            printf("User\n");
+        #endif
         PageDetails newUser;
         newUser.flags.flags = USER_FLAGS;
         newUser.flags.Execute_Disable = false;
@@ -19,7 +37,9 @@ void HandlePageFault(InterruptRegistersError* regs){
 
         alloc_page(&newUser);
     }else{
-        printf("Kernel\n");
+        #ifdef DEBUG
+            printf("Kernel\n");
+        #endif
         PageDetails newKernel;
         newKernel.flags.flags = KERNEL_FLAGS;
         newKernel.flags.Execute_Disable = false;
@@ -28,16 +48,7 @@ void HandlePageFault(InterruptRegistersError* regs){
 
         alloc_page(&newKernel);
     }
-
-    return;
-
-    asm(
-        "cli\n"
-        "1:\n\t"
-        "hlt\n"
-        "jmp 1b\n"
-        :::
-    );
+    SetTextColor(WHITE, BLACK);
 }
 
 void GeneralProtectionFault(InterruptRegistersError* regs){